+++ src/types/device.ts
@@ -1,0 +1,9 @@
+export type Device = {
+  width: number;
+  height: number;
+};
+
+export type AreaDevice = Device & {
+  offsetX: number;
+  position: number;
+};
