+++ src/types/index.ts
@@ -1,28 +1,3 @@
-export type Device = {
-  width: number;
-  height: number;
-};
-
-export type AreaDevice = Device & {
-  offsetX: number;
-  position: number;
-};
-
-export type Member = {
-  position: {
-    x: number;
-    y: number;
-  };
-  velocity: {
-    x: number;
-    y: number;
-  };
-  width: number;
-  height: number;
-  manager: string;
-  color?: string;
-};
-
-export type Members = {
-  [id: string]: Member;
-};
+export * from './device';
+export * from './member';
+export * from './player';
