+++ src/types/player.ts
@@ -1,0 +1,6 @@
+import { player } from "../enums";
+
+export type PlayerRole = {
+  role: player.Role,
+  properties: { interval: number } | null | null
+}
